Continue warning fixing

+++ include/microfmt/inspector/arm_exidx_decoder.hpp
@@ -81,12 +81,14 @@ private:
 
       // 00xxxxxx: vsp = vsp + (xxxxxx << 2) + 4
       if ((opcode & 0xC0) == 0x00) {
-        uint32_t offset = ((opcode & 0x3F) << 2) + 4;
+        uint32_t offset =
+            (static_cast<uint32_t>(opcode & 0x3F) << 2U) + 4U;
         io_sp += offset;
       }
       // 01xxxxxx: vsp = vsp - (xxxxxx << 2) - 4
       else if ((opcode & 0xC0) == 0x40) {
-        uint32_t offset = ((opcode & 0x3F) << 2) + 4;
+        uint32_t offset =
+            (static_cast<uint32_t>(opcode & 0x3F) << 2U) + 4U;
         io_sp -= offset;
       }
       // 1000iiii iiiiiiii: Pop integer registers (r4-r15)
@@ -95,13 +97,17 @@ private:
           break;
         uint8_t low_mask = bytes[idx++];
         uint16_t reg_mask =
-            (static_cast<uint16_t>(opcode & 0x0F) << 8) | low_mask;
+            static_cast<uint16_t>(
+                (static_cast<uint16_t>(opcode & 0x0F) << 8U) |
+                static_cast<uint16_t>(low_mask));
 
         for (int i = 0; i < 12; ++i) {
-          if ((reg_mask & (1 << i)) != 0) {
+          if ((reg_mask & (1U << static_cast<unsigned>(i))) != 0) {
             uint32_t val = 0;
             if (space.read_bytes(io_sp, &val, 4)) {
-              uint32_t dwarf_reg = dwarf::arm32::R4 + i;
+              uint32_t dwarf_reg =
+                  static_cast<uint32_t>(dwarf::arm32::R4) +
+                  static_cast<uint32_t>(i);
               if (!reg_ctx.write(dwarf_reg, val))
                 return false;
               io_sp += 4;
@@ -115,7 +121,9 @@ private:
         for (int i = 0; i <= count; ++i, io_sp += 8) {
           uint64_t val = 0;
           if (space.read_bytes(io_sp, &val, 8)) {
-            uint32_t dwarf_reg = dwarf::arm32::D0 + 8 + i;
+            uint32_t dwarf_reg =
+                static_cast<uint32_t>(dwarf::arm32::D0) + 8U +
+                static_cast<uint32_t>(i);
             if (!reg_ctx.write_raw(dwarf_reg, &val, 8))
               return false;
           }
